+++ java/benchmarks/src/main/java/javababushka/benchmarks/AsyncClient.java
@@ -1,0 +1,16 @@
+package javababushka.benchmarks;
+
+import java.util.concurrent.Future;
+
+public interface AsyncClient extends Client {
+
+  long DEFAULT_TIMEOUT = 1000;
+
+  Future<?> asyncSet(String key, String value);
+
+  Future<String> asyncGet(String key);
+
+  <T> T waitForResult(Future<T> future);
+
+  <T> T waitForResult(Future<T> future, long timeout);
+}
