+++ java/benchmarks/src/main/java/javababushka/benchmarks/Client.java
@@ -1,0 +1,13 @@
+package javababushka.benchmarks;
+
+import javababushka.benchmarks.utils.ConnectionSettings;
+
+public interface Client {
+  void connectToRedis();
+
+  void connectToRedis(ConnectionSettings connectionSettings);
+
+  default void closeConnection() {}
+
+  String getName();
+}
