+++ java/benchmarks/src/main/java/javababushka/benchmarks/SyncClient.java
@@ -1,0 +1,7 @@
+package javababushka.benchmarks;
+
+public interface SyncClient extends Client {
+  void set(String key, String value);
+
+  String get(String key);
+}
