+++ java/benchmarks/src/main/java/javababushka/benchmarks/utils/ChosenAction.java
@@ -1,0 +1,7 @@
+package javababushka.benchmarks.utils;
+
+public enum ChosenAction {
+  GET_NON_EXISTING,
+  GET_EXISTING,
+  SET
+}
