+++ java/benchmarks/src/main/java/javababushka/benchmarks/utils/ConnectionSettings.java
@@ -1,0 +1,13 @@
+package javababushka.benchmarks.utils;
+
+public class ConnectionSettings {
+  public String host;
+  public int port;
+  public boolean useSsl;
+
+  public ConnectionSettings(String host, int port, boolean useSsl) {
+    this.host = host;
+    this.port = port;
+    this.useSsl = useSsl;
+  }
+}
