+++ java/gradle.properties
