--- java/jabushka/benchmarks/src/main/java/javabushka/client/Client.java
@@ -1,13 +1,0 @@
-package javabushka.client;
-
-import javabushka.client.utils.ConnectionSettings;
-
-public interface Client {
-  void connectToRedis();
-
-  void connectToRedis(ConnectionSettings connectionSettings);
-
-  default void closeConnection() {}
-
-  String getName();
-}
