--- java/jabushka/benchmarks/src/main/java/javabushka/client/SyncClient.java
@@ -1,7 +1,0 @@
-package javabushka.client;
-
-public interface SyncClient extends Client {
-  void set(String key, String value);
-
-  String get(String key);
-}
