--- java/jabushka/benchmarks/src/main/java/javabushka/client/lettuce/LettuceClient.java
@@ -1,56 +1,0 @@
-/*
- * This Java source file was generated by the Gradle 'init' task.
- */
-package javabushka.client.lettuce;
-
-import io.lettuce.core.RedisClient;
-import io.lettuce.core.api.StatefulRedisConnection;
-import io.lettuce.core.api.sync.RedisStringCommands;
-import javabushka.client.SyncClient;
-import javabushka.client.utils.ConnectionSettings;
-
-public class LettuceClient implements SyncClient {
-
-  RedisClient client;
-  RedisStringCommands syncCommands;
-  StatefulRedisConnection<String, String> connection;
-
-  @Override
-  public void connectToRedis() {
-    connectToRedis(new ConnectionSettings("localhost", 6379, false));
-  }
-
-  @Override
-  public void connectToRedis(ConnectionSettings connectionSettings) {
-    client =
-        RedisClient.create(
-            String.format(
-                "%s://%s:%d",
-                connectionSettings.useSsl ? "rediss" : "redis",
-                connectionSettings.host,
-                connectionSettings.port));
-    connection = client.connect();
-    syncCommands = connection.sync();
-  }
-
-  @Override
-  public void set(String key, String value) {
-    syncCommands.set(key, value);
-  }
-
-  @Override
-  public String get(String key) {
-    return (String) syncCommands.get(key);
-  }
-
-  @Override
-  public void closeConnection() {
-    connection.close();
-    client.shutdown();
-  }
-
-  @Override
-  public String getName() {
-    return "Lettuce";
-  }
-}
