--- java/jabushka/benchmarks/src/main/java/javabushka/client/utils/ChosenAction.java
@@ -1,7 +1,0 @@
-package javabushka.client.utils;
-
-public enum ChosenAction {
-  GET_NON_EXISTING,
-  GET_EXISTING,
-  SET
-}
