--- java/jabushka/benchmarks/src/main/java/javabushka/client/utils/ConnectionSettings.java
@@ -1,13 +1,0 @@
-package javabushka.client.utils;
-
-public class ConnectionSettings {
-  public String host;
-  public int port;
-  public boolean useSsl;
-
-  public ConnectionSettings(String host, int port, boolean useSsl) {
-    this.host = host;
-    this.port = port;
-    this.useSsl = useSsl;
-  }
-}
